+++ src/components/shared/popup/KebabMenuButton.tsx
@@ -1,11 +1,10 @@
 import { theme } from "@/src/utils/theme";
 import { Box, Popper, Fade, SxProps } from "@mui/material";
 import Image from "next/image";
-import { Button } from "../button";
 import { useState } from "react";
 import { DeleteConfirmationModal } from "../../reflection-list/modal/DeleteConfirmationModal";
 import { red } from "@mui/material/colors";
-import PopupButton from "../button/PopupButton";
+import PopupButton from "./PopupButton";
 
 type KebabMenuButtonProps = {
   anchorEl: HTMLElement | null;
